server/requirements.txt
# FastAPI and Server
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# LangChain Core
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0

# LangChain Integrations
langchain-google-genai>=2.0.0
google-generativeai>=0.8.3

# Vector Store
faiss-cpu
# For GPU: faiss-gpu

# Database (if needed)
# sqlalchemy>=2.0.23
# psycopg2-binary>=2.9.9

# Utilities
python-dotenv>=1.0.0
httpx>=0.25.2

# Enhanced symbolic math for ReAct agent
sympy>=1.12

# Development
pytest>=7.4.3
pytest-asyncio>=0.21.1
